requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
ultralytics==8.3.0
pillow==10.4.0
python-multipart==0.0.12
opencv-python-headless==4.10.0.84
